requirements.txt
pymongo~=3.12.0
pydantic~=1.10.2
uvicorn~=0.20.0
fastapi~=0.95.1
pandas~=2.0.1
pyjwt~=2.4.0